requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
beautifulsoup4
requests
chardet